split routes index into smaller modules
