+++ src/app/routes/200/index.js
@@ -1,0 +1,10 @@
+import ErrorPage from '#pages/Error';
+import { errorPagePath } from '../regex';
+
+export default {
+  path: errorPagePath,
+  exact: true,
+  component: ErrorPage,
+  getInitialData: () => Promise.resolve({ status: 200 }),
+  pageType: 'error',
+};
