+++ src/app/routes/404/index.js
@@ -1,0 +1,7 @@
+import ErrorPage from '#pages/Error';
+
+export default {
+  component: ErrorPage,
+  getInitialData: () => Promise.resolve({ status: 404 }),
+  pageType: 'error',
+};
