+++ src/app/routes/article/index.js
@@ -1,0 +1,11 @@
+import getInitialData from '../getInitialData';
+import Article from '#pages/Article';
+import { articlePath } from '../regex';
+
+export default {
+  path: articlePath,
+  exact: true,
+  component: Article,
+  getInitialData,
+  pageType: 'article',
+};
