+++ src/app/routes/index/index.js
@@ -1,0 +1,11 @@
+import getInitialData from '../getInitialData';
+import FrontPage from '#pages/FrontPage';
+import { frontPagePath } from '../regex';
+
+export default {
+  path: frontPagePath,
+  exact: true,
+  component: FrontPage,
+  getInitialData,
+  pageType: 'frontPage',
+};
